@@ -1,6 +1,6 @@
 {
     "name": "dmyers/laravel-pages",
-    "description": "A simple static pages package for Laravel 4.",
+    "description": "A simple static pages package for Laravel 5.",
     "keywords": ["laravel", "pages", "static pages"],
     "license": "MIT",
     "authors": [
@@ -10,8 +10,8 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0",
-        "illuminate/support": "4.x"
+        "php": ">=5.4.0",
+        "illuminate/support": "~5.0"
     },
     "suggest": {
         "anahkiasen/flatten": "Add full page caching for speed."
