@@ -1,6 +1,6 @@
 <?php
 
-return array(
+return [
 	
 	/*
 	|--------------------------------------------------------------------------
@@ -44,7 +44,7 @@ return array(
 	|
 	*/
 	
-	'ignore_paths' => array(),
+	'ignore_paths' => [],
 	
 	/*
 	|--------------------------------------------------------------------------
@@ -57,4 +57,4 @@ return array(
 	
 	'cache' => false,
 	
-);
+];
